requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Pillow==10.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
